apiTest/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
cssutils==2.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==23.2.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
JPype1==1.5.0
jsonpath==0.82.2
loguru==0.7.2
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.2.0
playwright==1.42.0
pluggy==1.4.0
premailer==3.10.0
prestool==0.0.46
pyee==11.0.1
PyJWT==2.8.0
pymongo==4.6.2
PyMySQL==1.1.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-playwright==0.4.4
pytest-rerunfailures==14.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
radar==0.3
redis==5.0.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
yagmail==0.15.293
